udpate

+++ src/main/java/nodv/service/PostService.java
@@ -1,14 +1,17 @@
 package nodv.service;
 
+import nodv.exception.ForbiddenException;
 import nodv.exception.NotFoundException;
 import nodv.model.Post;
 import nodv.model.User;
 import nodv.repository.PostRepository;
 import nodv.security.TokenProvider;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,21 +25,17 @@ public class PostService {
     @Autowired
     TokenProvider tokenProvider;
 
-    public List<Post> findAll() {
-        return postRepository.findAll();
-    } // mongodb-method
-
-    public Post findById(String id) throws Exception {
+    // mongodb-method
+    public Post findById(String id) {
 
         Optional<Post> post = postRepository.findById(id);
         if (post.isEmpty()) {
-            throw new Exception("Post is not found");
+            throw new NotFoundException("Post is not found");
         }
         return post.get();
     }
 
-    public Post createPost(Post post, HttpServletRequest request) {
-        String userId = tokenProvider.getUserIdFromToken(tokenProvider.getJwtFromRequest(request));
+    public Post createPost(Post post, String userId) {
         User user = userService.findById(userId);
         post.setUserId(user.getId());
         post.setUser(user);
@@ -51,13 +50,37 @@ public class PostService {
         return postRepository.save(updatePost);
     }
 
-    public void deletePost(String id) {
+    public void deletePost(String id, String userId) {
         Optional<Post> post = postRepository.findById(id);
         if (post.isPresent()) {
-            postRepository.deleteById(id);
+            if (post.get().getUser().getId().equals(userId))
+                postRepository.deleteById(id);
+            else throw new ForbiddenException("You do not have permission to delete this post");
         } else {
             throw new NotFoundException("Post not found");
         }
     }
+
+    public void changePublish(String id, String userId, boolean isPublic) {
+        Post post = findById(id);
+        if (!post.getUser().getId().equals(userId))
+            throw new ForbiddenException("You do not have permission to update this post");
+        post.setIsPublish(isPublic);
+        postRepository.save(post);
+    }
+
+    public Page<Post> findAll(int page, int limit) {
+        Pageable pageable = PageRequest.of(page, limit);
+        return postRepository.findByIsPublishIsTrue(pageable);
+    }
+
+
+    public List<Post> findOwnedPost(String userId, String isPublish) {
+
+
+        if (isPublish == null) return postRepository.findByUserId(userId);
+        else return postRepository.findByUserIdAndIsPublish(userId, Boolean.valueOf(isPublish));
+    }
+
 
 }
